+++ CHANGELOG.md
@@ -9,7 +9,12 @@
 - [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.1.1)
 
 
-___
+### ⭐ 新特性
+
+- **buession-core：** DateTime 增加统计年、月天数
+
+
+---
 
 
 ## [2.1.0](https://github.com/buession/buessionframework/releases/tag/v2.1.0) (2022-08-07)